app/prospect_scoring_api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==0.28.1
python-dotenv==1.1.0
